Fix: Delay DB startup to ensure pins show
At present, there's an issue where Tomcat / NP comes up more slowly than
the Mongo database, which results in the map showing without any pins.

This 'fix' adds a sleep to the post run hook, which means that Tomcat
should have time to come up before the DB starts.

Of course, we should look into why this is happening and fix it more
elegantly, but for now this is the recommended approach.

Signed-off-by: Stephen Nelson-Smith stephen@atalanta-systems.com